Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> Before we do this, can we talk briefly about why we can't use properties here? We just made a few improvements to the underlying properties system, and my intuition right now is that the properties DB is the right thing here.

We thought about that earlier and it does have its merits, but we'd still need to special case quite a bit.  Property values are restricted to valid UTF8 characters, so we'd need to loosen that for the bit mask of requested_tags and the sequence number.  The enabled_cmdline goes way above the max length of a non-ro property (92 bytes), so we'd need to special case that as well.  We don't have any way to update only part of a property value either, so we'd be stuck doing RMW, whereas we don't need to with this.

It seems better to treat this as the special case that it is, rather than shoehorn it into properties.

Patch-set: 12
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>