Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > I'm rather hesitant here.  Do we have any precedent of mounting over a directory in /system?  I don't think so and I don't think it would be a good idea to start.
> 
> I guess this is the first case. I thought it would be nice if we can keep linker config under /system/etc as it is now, but if this is too big burden probably I should go with creating a new root-level directory like /linkerconfig.

I think that would be better overall.  /dev/linkerconfig may be a good option too, since we already have plenty of submounts in /dev, but I don't think it's too unreasonable for this to be a top level mount.  Definitely better than it being a submount of /system.  I do wonder in general if there're other things that may want the same behavior in the future rather than just the linkerconfig?  That may help us name it something different and make it even more useful as a top level mount.

> 
> > 
> > I also don't fully understand what the problem is with the runtime tests; both this CL and the bug are sparse on the description of the actual issue here.  How do the tests call chroot() and how does the chroot cause this issue?  Can either here or the bug have a link to the test in question or more details about what is going wrong?  I'd like to see if there are other ways to solve this issue.
> 
> I guess rpl@ can give you the better answer, but there are two scenarios (one for test, and other for update) which uses chrooted environment with system and some other directories. The problem from here is that two environment (chrooted and original) should use different linker configuration, but /dev directory is shared from both. It would be not easy to distinguish current environment from linker itself (it would be possible but too expensive in terms of time for linker), so I'd like to distinguish them by using different linker configuration but same path.

Understood.  Just wondering, are there other mounts that are not shared that could be reused?  /apex for example, since the dynamic linkerconfig does relate to apex?  Although I can understand the reasons not to want to use /apex for anything other than the bind mounted apexes as well.

Patch-set: 7
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>