Update patch set 10
Patch Set 10:

(1 comment)

It looks good to me as is, though I have a comment inline about whether or not we should do the mount() where we do.  Ideally, we move the mount to happen later, though this will work as is if we don't move it.

Patch-set: 10
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>