Update patch set 1
Patch Set 1:

Did you upload the wrong register file (regs.txt)? It's the registers for x86 not x86_64.

It's possible that the real problem is that somehow the code thinks that you are running as if you are x86 and not x86_64. That might cause all kinds of problems.

The code is in Regs.cpp::RemoteGet() for remote unwinds and Regs::CreateFromUcontext() for local unwinds. It's possible that might need to be modified for your use-case.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>