Update patch set 1
Patch Set 1:

Can you tell me what parts of the header file do we need?
Just the structure definition?  Or some of the #defines?

(ie. what portions of the header file can be deleted)

I'd just put the definitions straight in the .c file,
that way at least nothing else will start depending on the new .h.

We still have kernel quota2 hackery that might potentially depend/(re)uses this thing, cleaning that up has been on the todo list for more than a year now.

I'm not sure when/if we'll be able to get rid of this...
as a reminder we still need to support the 4.4 kernel in R (a full 4 years old already)...

Patch-set: 1
Reviewer: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>