Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Wondering if we could eliminate /cache as another source by instead using a bind mount of /mnt/scratch to /cache. Simplify a fair amount, and may work here as another strategy?

Because we have to support older devices we can't remove it yet, but I might be able to simplify things by using /cache as a backing store rather than a mount point. I'll put that on the list of things to try for this patch series.

Patch-set: 3
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>