Update patch set 7
Patch Set 7:

Marking the function pure (in the GCC attribute sense) would allow the compiler to not invalidate assumptions about global memory across the call. That internally it does modify memory (the tag initialization, as you mention) shouldn't matter because those modifications aren't visible to the outside world. Sure, it's a lie, but it's a lie that leads to better code.

Patch-set: 7