ImageManager returns errno.
IImageManager::CreateBackingImage and ZeroFillNewImage now returns
ImageManagerStatus, which is a wrapper of status_t error code and
provides more information about the source of error.

For compatibility with existing code, ImageManagerStatus may be
contextually converted to bool, and evaluates to true if error code is
OK and false otherwise.

In particular, the error code is -ENOSPC if disk space is full during
allocation / writing.

Bug: 138808058
Test: libsnapshot_test

Change-Id: I500a3d9f55a2d7e60438b5b4ae70f8b5fed162fa