init/service_parser: Add arguments `window' and `target' for `critical'
The critical services can now using the interface `critical
[window=<fatal crash window mins>] [target=<fatal reboot target>]` to
setup the timing window that more than 4 crashes in it is not tolerable,
and the reboot target of fatal handler as well.

Config `window=60 target=zygote-fatal' for all system-server services,
so the system will restart ifself when these UI sensitive processes
repeatedly crash.

Bug: 146818493
Change-Id: Ib2dc253616be6935ab9ab52184a1b6394665e813