Update patch set 1
Patch Set 1:

Is status_t actually what we want to use here?  It's better than errno since it's an explicitly returned value, but I'm always skeptical about these globally generic 'status' types.

How are callers going to use this value?  Do they make decisions based on how the functions failed or is it just for reporting?  If it's the former, can we create enum classes for these functions such that it's obvious to the caller and the function itself what error values should be returned?  If it's the latter, could we return strings here instead of trying to condense the information down to a single enum?

Also, are we otherwise using libutils here?  I guess we need it for some binder interaction?  That's an unfortunate library to have to bring in.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>