init: increase property service sndbuf size
There are some areas where init does not immediately return to its
main loop to handle property changed messages.  If too many are
generated, the sender will block, which may lead to deadlock if init
is waiting on the sender.  This specifically happens when init calls
to vold for mounting /data.

This increased buffer size is able to have >5000 pending property
changed messages waiting, which should be sufficient.

Bug: 146877356
Test: send >5000 property messages from a builtin command and observe
      when init starts blocking

Change-Id: I855de128b47fba35cfeaf9aea75c98bbd8ec4e91