init: handle properties in background of fs_mgr_mount_all
fs_mgr_mount_all may call out to vold, which may write
properties. Even though we now handle properties in a thread, so this
should be fine, there is still an issue if too many properties are set.

The property thread writes to a socket that is read from the main loop
of init.  This socket has a queue that can only handle ~200 property
set messages before it backs up.  Since it is non-blocking, it will
cause init to dead lock if vold sets more than this number of messages
before fs_mgr_mount_all finishes and init returns to its main loop.

Note that a better solution here would be to move all handling of
/data to vold and using `exec vdc ...` as a part of mount_all.  This
would allow init to return to its main loop and avoid this hack.

Bug: 146877356
Test: set 100k properties in fs_mgr_mount_all() without issue
Change-Id: Iaa8e322bbd06a0b4813224e8500bd42d6ebbef88