Don't call vdc markBootAttempt during userspace reboot
In case userspace reboot fails and fallback to hard reboot is triggered
we still want to try and activate any staged APEXes.

An installation of bad APEX using userspace reboot shouldn't trigger a soft
reboot loop because of the userspace reboot watchdog which will force a hard
reboot if boot hasn't completed in 2 minutes.

Test: builds
Bug: 135984674
Change-Id: If0a1340c810f13dcd9174fcadcbd44c86f604630