Update patch set 2
Patch Set 2:

> Patch Set 2: Presubmit-Ready+1
> 
> > Patch Set 2:
> > 
> > So, is the idea here that we don't trust userspace reboots? So we give it a free attempt at a userspace reboot which we don't count against the attempts, and the next reboot is always guaranteed to be a hard reboot? 
> 
> Either userspace reboot succeeds (and we commit checkpoint), or it fails and we fall back to hard reboot.
> 
> > Because if we can get another userspace reboot from this, wouldn't we loop until something else intervenes?
> 
> Technically something can issue an userspace reboot in the middle of the userspace reboot, but the at some point userspace reboot watchdog will kick in and force a hard reboot: https://cs.android.com/android/_/android/platform/system/core/+/931257683d189341d1a95b85ef9c50b8f2c48760:init/reboot.cpp;l=819;drc=018ddd7a6f0c176909af6c4f29d6cfea3b84c82b

You could just set the retry count to like 2 instead of 1, and then we're still covered in the case of repeated userspace reboots. If we don't expect that to be likely, it would have a similar affect. I'm a bit hesitant to be reliant on the watchdog if we don't need to be.

Patch-set: 2
Reviewer: Gerrit User 1045980 <1045980@85c56323-6fa9-3386-8a01-6480fb634889>