base: do not print filename or line number in log messages
StderrLogger and LogdLogger (only when FATAL) print a filename and
line number prefix to their log messages.  This information is not
typically useful, but moreover: libbase's logging will merge with
liblog's logging and liblog does not have this information.  To make
the log messages more symmetric, it is best to drop this information.

Also, fix a bug in the libbase unit tests, where the () in a regex was
not escaped properly.

Test: libbase unit tests pass
Change-Id: I18c0771e42ebfe39662b72e75d2b93c154f71707