liblog: check for loggability before formatting
Currently we do these actions in this order:
1) Format log the message
2) Get timestamp
3) Check if we're going to log the message

This change moves us to the the ideal order of:
1) Get timestamp
2) Check if we're going to log the message
3) Format the log message

This is the ideal order because:
1) should come first because we want to minimize drift from when
liblog gets called to log a message and when it determines the
timestamp.
2) should come second because we don't want to waste cycles formatting
the message if we're never going to log it anyway.

Also refactor some code along the way to make it easier to merge with
libbase.

Bug: 69935292
Bug: 119867234
Test: liblog-unit-tests
Change-Id: I8bfe898bcac55fd276968a439336305d3a10c4fd