Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> From some other threads: turns out libcutils from AOSP is shipping with some APEXes on Q, and therefore needs to be backwards compatible.  This is true about this CL, albeit that there may be some performance decrease since the system property calls in Q are not simply atomics, right?  It looks it to me at least.
> 
> Otherwise we may need to actually keep the binder code around for that use case, unfortunately.

Yes, that is the intention. I have been told that we never ship APEX that are older than the platform. So we need to support [Q atrace + new libcutils] as well as [new atrace + new libcutils], which we do with this CL. The old atrace sadly sets the cmdline property after the tags one, so we cannot remove the sysprop handler in this CL (which would not be a great performance improvement anyway, as the problem seem to be *other* sysprop handlers, which would anyway still get invoked).

Patch-set: 11
Reviewer: Gerrit User 1268571 <1268571@85c56323-6fa9-3386-8a01-6480fb634889>