Rename libstatssocket to libstatspush
We rename this library for two reasons: (a) writing to the socket is an
implementation detail and (b) many clients of this library will not
log events to the socket.

Note that this CL does not actually remove libstatssocket. Instead, it
only creates a copy called libstatspush. This is carefully done to
prevent automerger issues when AOSP merges into google master.
Specifically, there are libraries within google master that depend on
libstatssocket and will be broken by its removal. We will remove
libstatssocket only once libstatspush has been merged into google master
and when the aforementioned clients depend on libstatspush and not
libstatssocket.

Test: m -j
Bug: 146377827
Change-Id: Idcd8d0afacf586a07312ba135c2259f30d56cf16