Update patch set 11
Patch Set 11:

I was hoping to allow others to chime in still with my +1.  If no one does I can +2 tomorrow.

Did we ever get an answer for "How to deprecate ATRACE_SHMEM?"  Do we have it just in case this doesn't work that we can disable the flag instead of reverting the CL?  Perhaps add a comment //TODO: Delete by XXX date so we can put a fixed amount of time on its short-livedness?

Patch-set: 11
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>