Make AvbHandle able to verify DSU after SwitchRoot
The bug happened when the device installed a DSU package that
contains more than one partition, e.g. product and system, and
the system gets mounted before others. Because AvbHandle::LoadAndVerifyVbmeta
reads files under /avb which are only available in the 1st stage
ramdisk. If the SwitchRoot happened before any of the DSU partition get
mounted the AvbHandle::LoadAndVerifyVbmeta will fail to read the allowed
public key and thus fail to boot.

Bug: 147585411
Test: adb shell am start-activity \
    -n com.android.dynsystem/com.android.dynsystem.VerificationActivity \
    -a android.os.image.action.START_INSTALL \
    -d file:///storage/emulated/0/Download/dsu.zip \
    --el KEY_USERDATA_SIZE 8589934592

Change-Id: I6f0c3142944480628850bcf897d44197f4f350fd