libprocessgroup: add HasFile() method to CgroupController
It can be useful to check the presence of a particular file in a cgroup
controller. For instance, util_clamp is an extension of the CPU
controller, so checking its presence requires not only to verify that
/dev/cpuctl exists, but also that cpu.uclamp.* files are present in
sub-groups.

To ease the implementation of such checks, extend the CgroupController
class with a HasFile() method checking the presence of arbitrary files.

Change-Id: I09806e668c9d08256efdadb783dcf3ac596b8d77
Signed-off-by: Quentin Perret <qperret@google.com>