Update patch set 1
Patch Set 1:

That's rather quick compared to the start up time of a process.  Do you need this to be done synchronously in init?  Do you need it to be completed by a certain point?

If not, then it's better as a separate process, to have clear lines of ownership and separating the features.

If you need it to be completed by a certain point, we may still want it to be a different process even.  We can always launch it early then have init wait for its completion.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>