Update patch set 6
Patch Set 6:

> > It looks like we need to add an implementation for Transport::Reset() for the ClientUsbTransport class as well.  Can you add this?
> 
> How about this, i removed virtual Reset() in Transport so that we don't need empty implementation for tcp, udp and ClientUsbTransport. TransportSniffer only need to call UsbTransport's Reset() when dynamic_cast is true.
> Thank you.

We don't have RTTI enabled, so dynamic_cast<> is not a good option.  It would be better to add the Reset() to Transfer like you originally did; we just need to make sure that the all of the subclasses implement it.

Thanks!

Patch-set: 6
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>