Update patch set 1
Patch Set 1:

I'm not quite sure what the plan is here. (I was on vacation last week, and didn't get a chance to chime in with the earlier CL.)

Until recently we did this in stdatomic.h, which implicitly included <atomic> instead, if invoked from C++. This is also the proposal in wg21.link/p0943, which is slowly winding its way through the C++ committee, and currently looks as though it will probably be in C++23. Was that effectively removed by aosp/1209251? Or were host builds not doing this consistently anyway?

It seems to me that if we can do this in stdatomic.h we should. If we can't do it there, we should probably put the full list of using declarations from p0943, together with the <stdatomic.h> or <atomic>, in a new header that can be included instead of <stadatomic.h>, and does the right thing.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>