Update patch set 1
Patch Set 1:

I don't think the behavior change is that big. It will result in debuggerd displaying the registers in a slightly different order, but that's not too bad. I verified that the debuggerd unit tests still pass, so I'm fine with this.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>