libprocessgroup: Prevent aggregate profiles from referencing stale profiles
Currently when vendor overrides a profile the profile object is being
replaced with a new one. However the old profile might have been referenced
by an aggregate profile and with such profile replacement the aggregate
profile is left referencing a stale object. Fix this by replacing the
content of the old profile with the content from the new one instead of
replacing the object itself.

Bug: 148311066
Test: override profiles referenced in aggregate profile and verify
Test: correct replacement
Change-Id: Iabddbf3580455e5263fedad6665cf52fb323e50a
Signed-off-by: Suren Baghdasaryan <surenb@google.com>