Add libsnapshot_test to presubmit
Create libsnapshot_test's twin vts_libsnapshot_test that will be part of
the VTS tests suite.
Also add libsnapshot test to the TEST_MAPPING's presubmit of fs_mgr to
run the test on TH whenever changes are applied to the fs_mgr contents.
Some of the tests are temporarily disabled due to a WIP bug.

Test: TH
Bug: 142513589
Change-Id: I2a283645b433ed1fd2b08d1c4919cac26d19cf16
Signed-off-by: Alessio Balsini <balsini@google.com>