Update patch set 1
Patch Set 1:

> my point is that i think people think they're getting more from symbols_file than they actually are. i suspect that everyone who's using symbols_file should also be using version_script, or they're likely falling into this same trap.

IMO, a big downside of requiring version_script is that every private symbols also need to be explicitly listed, because otherwise, they won't be available to other libraries even in the **same** module. For example, "local: *;" in many of the existing map.txt files will hide all (except the ones explicitly listed) symbols from outside of the library regardless of whether the client is from the same module or not. To resolve the issue, the lib author should explicitly associate a private version (having _PRIVATE or _PLATFORM suffix) to the each of the private symbols. This however is very challenging when the lib has to export C++ symbols to its peers in the same module. I wonder if that's what we want. (I do agree that it's safer though)

Patch-set: 1