Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Looks good to me.  If this documented anywhere that we'll need to change as well?

It seems not widely used so probably we're good.
https://cs.android.com/search?q=FB_RESPONSE_SZ

> 
> Is 128 the largest that we can envision?  

This is mostly for fingerprint but I'm not sure how large it could be.
We're adding VTS for that so eventually we'll know.
https://android-review.googlesource.com/c/platform/test/vts/+/1226735


> Is there any disadvantage in making it even larger?
Yup, if not much concern, we can probably make it even larger.

Let's see if David/Hridya has any thoughts or not.

Patch-set: 1