Update patch set 2
Patch Set 2: Code-Review+2

> That said, the C++20 std::format() proposal doesn't seem to support named arguments, so maybe it's fine to remove this test and not support them with Result<>?  Users can always fall back to using fmtlib directly if it's utterly required.

Yeah, this would be another opinionated decision, but I feel fine with it. Named arguments are not popular and as you said there is a way to use it if really needed.

> On a related topic, there used to be a 'std' branch on github for fmtlib that implemented the standard as is.  It was merged into the master branch, which I guess should support std::format() directly now, but I simply don't see the configuration or any mention of std::format() in it.  It seems like a better idea if we can use the std::format() subset of fmtlib now instead of having a harder switch later.

Yeah, if the standard is expected to not support it, we'd better not use it.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2