Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > should we avoid divergence from the standard, or what is the verdict there?
> > > 
> > > It seems like we're burning that bridge already when removing the operators.
> > 
> > Perhaps I could have annotated the places where we diverge from P0323R9.
> > Or we could wait to see the outcome of the WG21 meeting next week:
> > 
> >   https://isocpp.org/std/meetings-and-participation/upcoming-meetings
> 
> Not exactly. The other change meant that we would still be compatible when std::expected ships.
> This change means that we could not switch to the standard w/o making client code changes.

All we need to do is mechanically replace ok() with has_value()... which is trivial compared to the kind of changes that could be requested by the committee.

An alternative I considered was defining ok() in Result, but that would turn Result into a derived class, which would require forwarding constructors and operators to convert it seamlessly to base::expected.

Let's spare ourselves the burden of compatibility until we know whether std::expected will actually land in C++20, and how much it will change.

Patch-set: 1
Reviewer: Gerrit User 1273909 <1273909@85c56323-6fa9-3386-8a01-6480fb634889>