ueventd: Remove strict loading of modules on modalias uevent
After finding a device in a system the kernel will emit a modalias
uevent in-case there is a loadable module for that device. This
is often not the case as many device drivers are built-in. Loading
a module for is not required every time, so drop the strict flag
to reduce the number of console warnings.

Test: Check boot log for less "LoadWithAliases was unable to load" warns
Change-Id: I6ea6e7667d42d1d15fc9b4d2ba8c27a5dfc7f961
Signed-off-by: Andrew F. Davis <afd@ti.com>