"track-app" service showing debuggable/profileable apps
Add a "track-app" service in adbd. For every debuggable or
profileable-from-shell process, ART sends related info to
adbd and adbd surfaces the info through the "track-app"
service.

The output format of "track-app" is very similar to the
existing "track-jdwp" service. For example, a process is
shown as string "3529-d0p1\n" where
    3529 is the pid,
    d0 means it's no debuggable,
    p1 means it's profileable-from-shell.

Bug: 149050485
Test: N/A

Change-Id: Id1f1a920e1afc148c7e4d2add790baab796178e1