"track-app" service showing debuggable/profileable apps
Add a "track-app" service in adbd. For every debuggable or
profileable-from-shell process, ART sends related info to
adbd and adbd surfaces the info through the "track-app"
service.

The output format of "track-app" is a line summarizing
the number of reported processes, followed by a protobuf
message in human readable form. For example,

2 debuggable/profileable app processes reported.
process {
  pid: 9689
  debuggable: true
  profileable: true
}
process {
  pid: 9724
  profileable: true
}

Bug: 149050485
Test: TODO

Change-Id: Id1f1a920e1afc148c7e4d2add790baab796178e1